refactor : set naming convention of files to Pascal and follow app folder containing files conventios
