+++ App.tsx
@@ -1,5 +1,5 @@
 import { useDeviceOrientation } from "@react-native-community/hooks";
-import Welcome from "./screens/welcome";
+import WelcomeScreen from './app/screens/WelcomeScreen';
 export default function App() {
   const { landscape } = useDeviceOrientation();
   // Information about dimenstions of screen
@@ -8,7 +8,7 @@ export default function App() {
   const image = { uri: "../assets/background.jpg" };
   return (
     <>
-      <Welcome title='Welcome Screen' orintation={landscape} />
+      <WelcomeScreen title='Welcome Screen' orintation={landscape} />
     </>
   );
 }
