+++ app/assets/adaptive-icon.png
