+++ app/assets/chair.jpg
