+++ app/components/RegisterButton.tsx
